remove unrequired files
