+++ api/v1/server/server.go
@@ -19,7 +19,7 @@ func ApplyRoutes(r *gin.RouterGroup) {
 		g.GET("", readServer)
 		g.PATCH("", updateServer)
 		g.GET("/config", configServer)
-		g.GET("/status", GetStatus)
+		
 	}
 }
 
@@ -98,26 +98,4 @@ func configServer(c *gin.Context) {
 	// return config as txt file
 	c.Header("Content-Disposition", "attachment; filename="+os.Getenv("WG_INTERFACE_NAME")+"")
 	c.Data(http.StatusOK, "application/config", configData)
-}
-
-// swagger:route GET /server/status Server statusServer
-//
-// Get Server status
-//
-// Retrieves the server  status details.
-// responses:
-//  200: serverStatusResponse
-//  400: badRequestResponse
-//	401: unauthorizedResponse
-//  500: serverErrorResponse
-func GetStatus(c *gin.Context) {
-	status_data, err := core.GetServerStatus()
-	if err != nil {
-		log.WithFields(util.StandardFields).Error("Failed to get server status")
-		response := core.MakeErrorResponse(500, err.Error(), nil, nil, nil)
-		c.JSON(http.StatusInternalServerError, response)
-		return
-	}
-
-	c.JSON(http.StatusOK, status_data)
 }
