+++ api/v1/v1.go
@@ -3,6 +3,7 @@ package v1
 import (
 	"github.com/TheLazarusNetwork/erebrus/api/v1/client"
 	"github.com/TheLazarusNetwork/erebrus/api/v1/server"
+	"github.com/TheLazarusNetwork/erebrus/api/v1/status"
 
 	"github.com/gin-gonic/gin"
 )
@@ -13,5 +14,6 @@ func ApplyRoutes(r *gin.RouterGroup) {
 	{
 		client.ApplyRoutes(v1)
 		server.ApplyRoutes(v1)
+		status.ApplyRoutes(v1)
 	}
 }
