+++ core/server.go
@@ -20,8 +20,6 @@ import (
 	"golang.zx2c4.com/wireguard/wgctrl/wgtypes"
 )
 
-var PublicKey string
-
 // ReadServer object, create default one
 func ReadServer() (*model.Server, error) {
 	if !util.FileExists(filepath.Join(os.Getenv("WG_CONF_DIR"), "server.json")) {
@@ -33,7 +31,6 @@ func ReadServer() (*model.Server, error) {
 		}
 		server.PrivateKey = key.String()
 		server.PublicKey = key.PublicKey().String()
-		PublicKey = server.PublicKey
 		server.Endpoint = os.Getenv("WG_ENDPOINT_HOST")
 		listenPort, _ := strconv.ParseInt(os.Getenv("WG_ENDPOINT_PORT"), 10, 32)
 
@@ -210,7 +207,7 @@ func GetServerStatus() (*model.Status, error) {
 	serverStatus, err := storage.Deserialize("server.json")
 
 	if err != nil {
-		log.WithFields(util.StandardFields).Error(err)
+		log.WithFields(util.StandardFields).Fatal(err)
 	} else {
 		var server model.Server
 		bodybytes, _ := json.Marshal(serverStatus)
