--- wg0.conf
@@ -1,10 +1,0 @@
-# Updated: 858889900 / Created: 858889900
-[Interface]
-Address = 10.0.0.1/24
-ListenPort = 51820
-PrivateKey = cHt4HCu6I8cJCFFqMdDhm4lXmRuCoQjOKh0nBYalY3g=
-
-PreUp = echo WireGuard PreUp
-PostUp = iptables -A FORWARD -i %i -j ACCEPT; iptables -A FORWARD -o %i -j ACCEPT; iptables -t nat -A POSTROUTING -o eth0 -j MASQUERADE
-PreDown = echo WireGuard PreDown
-PostDown = iptables -D FORWARD -i %i -j ACCEPT; iptables -D FORWARD -o %i -j ACCEPT; iptables -t nat -D POSTROUTING -o eth0 -j MASQUERADE
